+++ DomainLayer/Entities/UserAccount.cs
@@ -10,7 +10,7 @@ namespace DomainLayer.Entities
 
         public string Avatar { get; set; }
 
-        public int? ReferredByUserId { get; set; }
+        public long? ReferredByUserId { get; set; }
 
         public string UserName { get; set; }
